Adopt Cdecl calling convention
Egor reports that the only way to get Windows to run on 32 bit mode
with Urho when passing Quaternions and Vectors is to use Cdecl.  This
means that now this can run on 32 and 64 bits.

This seems to also work on the Mac, with the very simple sample that I
used.